ActiveRecord - Organizing all error classes
