+++ activerecord/lib/active_record/connection_adapters/abstract/schema_statements.rb
@@ -1026,7 +1026,7 @@ module ActiveRecord
         inserting = (versions - migrated).select { |v| v < version }
         if inserting.any?
           if (duplicate = inserting.detect { |v| inserting.count(v) > 1 })
-            raise "Duplicate migration #{duplicate}. Please renumber your migrations to resolve the conflict."
+            raise DuplicateMigrationError, duplicate
           end
           if supports_multi_insert?
             execute insert_versions_sql(inserting)
@@ -1060,7 +1060,7 @@ module ActiveRecord
             if (0..6) === precision
               column_type_sql << "(#{precision})"
             else
-              raise(ActiveRecordError, "No #{native[:name]} type has precision of #{precision}. The allowed range of precision is from 0 to 6")
+              raise ArgumentError, "No #{native[:name]} type has precision of #{precision}. The allowed range of precision is from 0 to 6"
             end
           elsif (type != :primary_key) && (limit ||= native.is_a?(Hash) && native[:limit])
             column_type_sql << "(#{limit})"
