+++ activerecord/lib/active_record/connection_adapters/postgresql/schema_statements.rb
@@ -539,21 +539,21 @@ module ActiveRecord
               # The hard limit is 1GB, because of a 32-bit size field, and TOAST.
               case limit
               when nil, 0..0x3fffffff; super(type)
-              else raise(ActiveRecordError, "No binary type has byte size #{limit}.")
+              else raise(AdapterTypeInconsistentWithByteLength, "No binary type has byte size #{limit}.")
               end
             when "text"
               # PostgreSQL doesn't support limits on text columns.
               # The hard limit is 1GB, according to section 8.3 in the manual.
               case limit
               when nil, 0..0x3fffffff; super(type)
-              else raise(ActiveRecordError, "The limit on text can be at most 1GB - 1byte.")
+              else raise(AdapterTypeInconsistentWithByteLength, "The limit on text can be at most 1GB - 1byte.")
               end
             when "integer"
               case limit
               when 1, 2; "smallint"
               when nil, 3, 4; "integer"
               when 5..8; "bigint"
-              else raise(ActiveRecordError, "No integer type has byte size #{limit}. Use a numeric with scale 0 instead.")
+              else raise(AdapterTypeInconsistentWithByteLength, "No integer type has byte size #{limit}. Use a numeric with scale 0 instead.")
               end
             else
               super
