+++ activerecord/lib/active_record/errors/active_record_error.rb
@@ -1,0 +1,6 @@
+module ActiveRecord
+  # Generic Active Record exception class.
+  # Please, raise a subclass when contributing.
+  class ActiveRecordError < StandardError
+  end
+end
