+++ activerecord/lib/active_record/fixtures.rb
@@ -8,9 +8,6 @@ require "active_record/fixture_set/file"
 require "active_record/errors"
 
 module ActiveRecord
-  class FixtureClassNotFound < ActiveRecord::ActiveRecordError #:nodoc:
-  end
-
   # \Fixtures are a way of organizing data that you want to test against; in short, sample data.
   #
   # They are stored in YAML files, one file per model, which are placed in the directory
