+++ activerecord/test/cases/adapters/postgresql/bytea_test.rb
@@ -33,7 +33,7 @@ class PostgresqlByteaTest < ActiveRecord::PostgreSQLTestCase
 
   def test_binary_columns_are_limitless_the_upper_limit_is_one_GB
     assert_equal "bytea", @connection.type_to_sql(:binary, limit: 100_000)
-    assert_raise ActiveRecord::ActiveRecordError do
+    assert_raise ActiveRecord::AdapterTypeInconsistentWithByteLength do
       @connection.type_to_sql(:binary, limit: 4294967295)
     end
   end
