+++ activerecord/test/cases/time_precision_test.rb
@@ -35,7 +35,7 @@ if subsecond_precision_supported?
     end
 
     def test_invalid_time_precision_raises_error
-      assert_raises ActiveRecord::ActiveRecordError do
+      assert_raises ArgumentError do
         @connection.create_table(:foos, force: true) do |t|
           t.time :start,  precision: 7
           t.time :finish, precision: 7
